requirements.txt
# Flask web framework
Flask
flask-cors
flask-sqlalchemy
Flask-Login>=0.6.3

# Flask-WTF for CSRF protection
Flask-WTF

# Cryptography for file encryption/decryption
cryptography

# Requests for HTTP file download
requests

# Watchdog for real-time folder monitoring
watchdog

# psutil for process management
psutil

# pandas, matplotlib, seaborn, numpy, scipy for data analysis (used in data_analysis.py)
pandas
matplotlib
seaborn
numpy==1.26.3
scipy==1.13.0
scikit-learn==1.6.1

# Optional: clamd for ClamAV integration on Linux/macOS
clamd>=1.0

# For automated testing
pytest>=7.0

# For loading .env files
python-dotenv>=1.0

# For Windows shortcut creation (Desktop shortcut)
pywin32

# Windows-specific packages for admin privileges
pywin32-ctypes>=0.2.0
wmi>=1.5.1

# Werkzeug for secure_filename (if used directly)
Werkzeug>=2.0

# Archive extraction support
rarfile>=4.0

# YARA for signature scanning
yara-python>=4.0

dnslib>=0.9.26

# For secure password hashing
bcrypt>=4.0

# For TOTP-based 2FA
pyotp>=2.9
rarfile>=1.0
py7zr>=0.21.1

# DNS resolution for Project Honey Pot
dnslib>=0.9.22

# For DNS-based reputation services
dnspython>=2.0

# Machine Learning dependencies
# Note: If installing scikit-learn from source, please build the package first.
# See: https://scikit-learn.org/dev/developers/advanced_installation.html#building-from-source
# YARA for advanced malware scanning
scikit-learn>=1.3.0
joblib>=1.3.0
requests>=2.31.0
psutil>=5.9.0

# Network monitoring and analysis
flask-limiter==3.12
limits==5.1.0
ordered-set==4.1.0
rich==13.9.4
pygments==2.19.1
typing_extensions==4.13.2
wrapt==1.17.2
markdown-it-py==3.0.0
mdurl==0.1.2
yara-python
python-magic>=0.4.27
pefile>=2022.5.30

# For VirusTotal API integration
# (requests already included)

# For logging rotation
# (logging is stdlib, but logging.handlers is used)

# For Windows Firewall (no pip package, uses subprocess)

# For inspecting file metadata
pefile>=2023.2.7
redis>=5.0.1
